+++ pkg/controllers/deprovisioning/consolidation.go
@@ -63,7 +63,11 @@ func makeConsolidation(clock clock.Clock, cluster *state.Cluster, kubeClient cli
 }
 
 // consolidationTTL is the TTL between creating a consolidation command and validating that it still works.
-const consolidationTTL = 15 * time.Second
+const (
+	consolidationTTL = 15 * time.Second
+	consolidationMethod = "Consolidation"
+)
+
 
 // string is the string representation of the deprovisioner
 func (c *consolidation) String() string {
