+++ pkg/controllers/provisioning/scheduling/topology.go
@@ -22,11 +22,10 @@ import (
 	"github.com/aws/karpenter-core/pkg/controllers/state"
 	"github.com/aws/karpenter-core/pkg/scheduling"
 
+	utilsscheduling "github.com/aws/karpenter-core/pkg/utils/scheduling"
 	"go.uber.org/multierr"
 	v1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/labels"
-	"k8s.io/apimachinery/pkg/selection"
 	"k8s.io/apimachinery/pkg/types"
 	utilsets "k8s.io/apimachinery/pkg/util/sets"
 	"sigs.k8s.io/controller-runtime/pkg/client"
@@ -235,7 +234,7 @@ func (t *Topology) countDomains(ctx context.Context, tg *TopologyGroup) error {
 	// simultaneously)
 	var pods []v1.Pod
 	for _, ns := range tg.namespaces.UnsortedList() {
-		if err := t.kubeClient.List(ctx, podList, TopologyListOptions(ns, tg.selector)); err != nil {
+		if err := t.kubeClient.List(ctx, podList, utilsscheduling.TopologyListOptions(ns, tg.selector)); err != nil {
 			return fmt.Errorf("listing pods, %w", err)
 		}
 		pods = append(pods, podList.Items...)
@@ -361,44 +360,6 @@ func (t *Topology) getMatchingTopologies(p *v1.Pod, requirements scheduling.Requ
 		}
 	}
 	return matchingTopologies
-}
-
-func TopologyListOptions(namespace string, labelSelector *metav1.LabelSelector) *client.ListOptions {
-	selector := labels.Everything()
-	if labelSelector == nil {
-		return &client.ListOptions{Namespace: namespace, LabelSelector: selector}
-	}
-	for key, value := range labelSelector.MatchLabels {
-		requirement, err := labels.NewRequirement(key, selection.Equals, []string{value})
-		if err != nil {
-			return &client.ListOptions{Namespace: namespace, LabelSelector: labels.Nothing()}
-		}
-		selector = selector.Add(*requirement)
-	}
-	for _, expression := range labelSelector.MatchExpressions {
-		requirement, err := labels.NewRequirement(expression.Key, mapOperator(expression.Operator), expression.Values)
-		if err != nil {
-			return &client.ListOptions{Namespace: namespace, LabelSelector: labels.Nothing()}
-		}
-		selector = selector.Add(*requirement)
-	}
-	return &client.ListOptions{Namespace: namespace, LabelSelector: selector}
-}
-
-func mapOperator(operator metav1.LabelSelectorOperator) selection.Operator {
-	switch operator {
-	case metav1.LabelSelectorOpIn:
-		return selection.In
-	case metav1.LabelSelectorOpNotIn:
-		return selection.NotIn
-	case metav1.LabelSelectorOpExists:
-		return selection.Exists
-	case metav1.LabelSelectorOpDoesNotExist:
-		return selection.DoesNotExist
-	}
-	// this shouldn't occur as we cover all valid cases of LabelSelectorOperator that the API allows.  If it still
-	// does occur somehow we'll panic just later when the requirement throws an error.,
-	return ""
 }
 
 func IgnoredForTopology(p *v1.Pod) bool {
